# Add const iteration support to `Dictionary`