Based on the assumption in #103999, we are able to bring `ConstIterator` to `Dictionary` without breaking the `read_only` property. Since the inner `HashMap` already has a nice iterator implementation, I just export it with alias to keep the change minimal.

Main usecase of this is to replace `Dictionary::get_key_list()`. Previously if we want to iterate a `Dictionary` we have to write:
```C++
List<Variant> keys;
dict.get_key_list(&keys);
for (const Variant &key : keys) {
    const Variant &value = dict[key];
    ......
}
```

This pattern includes unnecessary heap allocation (the `List`), redundant iteration (populate keys in `get_key_list()`) and an extra hashmap lookup (if you need both key and value).

With `ConstIterator` we can now write:
```C++
for (const KeyValue<Variant, Variant> &kv : dict) {
    const Variant &key = kv.key;
    const Variant &value = kv.value;
    ......
}
```

No heap allocation, scan exactly one time and no hashmap lookup.

I measure performance diff between the two pattern with [MRP from #100333](https://github.com/godotengine/godot/pull/100333#issuecomment-2583278772) (Note this workload is constructed intentionally to be bottlenecked at `Dictionary` iteration, `Dictionary` stringify would be a better showcase but unfortunately it is now bottlenecked by `String` concat)

Master:
```bash
729 ms
732 ms
729 ms
729 ms
727 ms
729 ms
725 ms
725 ms
720 ms
714 ms
average: 725 ms
```

This PR:
```bash
271 ms
270 ms
273 ms
271 ms
271 ms
272 ms
271 ms
273 ms
271 ms
272 ms
average: 271 ms
```

Callers who sort the key list or modify values in the loop are kept untouched.