@Repiteo 2025-03-13
Might be jumping the gun a bit here, but I'm confident enough about dictionaries to send this as-is.

Thanks!